Api Wilayah for address

+++ app/category/controller.js
@@ -2,6 +2,15 @@ const Category = require("./model");
 
 async function store(req, res, next) {
   try {
+    let policy = policyFor(req.user);
+
+    if (!policy.can("create", "Category")) {
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk membuat produk`,
+      });
+    }
+
     let payload = req.body;
     let category = new Category(payload);
 
@@ -23,6 +32,15 @@ async function store(req, res, next) {
 
 async function update(req, res, next) {
   try {
+    let policy = policyFor(req.user);
+
+    if (!policy.can("update", "Category")) {
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk mengupdate category`,
+      });
+    }
+
     let payload = req.body;
     let category = await Category.findOneAndUpdate(
       { _id: req.params.id },
@@ -46,6 +64,15 @@ async function update(req, res, next) {
 
 async function destroy(req, res, next) {
   try {
+    let policy = policyFor(req.user);
+
+    if (!policy.can("delete", "Category")) {
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk menghapus category`,
+      });
+    }
+
     let deleted = await Category.findOneAndDelete({ _id: req.params.id });
     return res.json(deleted);
   } catch (err) {
