+++ app/tag/controller.js
@@ -3,6 +3,14 @@ const Tag = require("./model");
 
 async function store(req, res, next) {
   try {
+    let policy = policyFor(req.user);
+    if (!policy.can("create", "Tag")) {
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk membuat tag`,
+      });
+    }
+
     let playload = req.body;
     let tag = new Tag(playload);
 
@@ -23,6 +31,13 @@ async function store(req, res, next) {
 
 async function update(req, res, next) {
   try {
+    let policy = policyFor(req.user);
+    if (!policy.can("update", "Tag")) {
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk mengupdate tag`,
+      });
+    }
     let payload = req.body;
 
     let tag = await Tag.findOneAndUpdate({ _id: req.params.id }, payload, {
@@ -46,6 +61,14 @@ async function update(req, res, next) {
 
 async function destroy(req, res, next) {
   try {
+    let policy = policyFor(req.user);
+    if (!policy.can("delete", "Tag")) {
+      return res.json({
+        error: 1,
+        message: `Anda tidak memiliki akses untuk delete tag`,
+      });
+    }
+
     let tag = await Tag.findOneAndDelete({ _id: req.params.id });
     return res.json(tag);
   } catch (err) {
